Publish the characteristic impedance from the medium, not from the tube
rho c is a property of the medium. It was published by materials because the
impedance tube was the first thing to need it, and room.steady_field,
emission.sound_power_in_duct, the layered and porous models and the tube itself
all reached into that package for a product of two numbers.

The addition and the withdrawal are one commit because one-name-one-owner keys
on the spelling: published by both packages, even for a single commit in
between, is what the taxonomy gate refuses.

The arguments keep density and speed_of_sound. The unit goes in a parameter
name where two units are in play and a caller can hand over the wrong one,
which is why the temperatures say _c and the pressures _pa. A density here is
kilograms per cubic metre everywhere; a suffix would buy nothing and would make
this function differ from the two hundred parameters already named that way.
