+++ src/phonometry/materials/__init__.py
@@ -76,7 +76,6 @@ from .absorbers import (
     attenuation_from_alpha,
     biot_surface_impedance,
     biot_waves,
-    characteristic_impedance,
     critical_coupling_design,
     decoupling_frequency,
     delany_bazley,
@@ -292,7 +291,6 @@ __all__ = [
     "base_plate_scattering",
     "biot_surface_impedance",
     "biot_waves",
-    "characteristic_impedance",
     "check_base_plate_scattering",
     "check_spot_frequency_range",
     "critical_coupling_design",
