+++ src/phonometry/materials/absorbers/__init__.py
@@ -42,7 +42,6 @@ from .impedance_tube import (
     absorption_from_reflection,
     air_density_iso10534,
     apply_mic_calibration,
-    characteristic_impedance,
     hydraulic_diameter,
     mic_calibration_factor,
     normalized_surface_admittance,
@@ -185,7 +184,6 @@ __all__ = [
     "attenuation_from_alpha",
     "biot_surface_impedance",
     "biot_waves",
-    "characteristic_impedance",
     "critical_coupling_design",
     "decoupling_frequency",
     "delany_bazley",
